pyproject.toml
[project]
name = "langchain-rag"
version = "0.1.0"
description = "AI Engineer's Guide to LangChain. Full course on LangChain with OpenAI."
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-core>=0.3.15",
    "langchain-openai>=0.2.5",
    "langsmith>=0.1.139",
    "langchain-community>=0.3.5",
    "langchain>=0.3.7",
    "docarray>=0.40.0",
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "google-search-results>=2.4.2",
    "python-dotenv>=1.0.1",
    "scikit-image>=0.25.2",
    "matplotlib>=3.10.3",
    "pypdf>=5.6.0",
    "chromadb>=1.0.12",
    "pandas>=2.3.0",
    "scikit-learn>=1.7.0",
    "faiss-cpu>=1.11.0",
    "streamlit>=1.45.1",
    "dspy>=2.6.27",
    "ragas>=0.2.15",
    "rapidfuzz>=3.13.0",
]
